Refactoring mangata helpers
